Refactored precedence rules for partitioned datasets
Still have to implement all the functions to properly allocate datasets on like half of use cases, but this is a checkpoint to see if the rules I proposed would handle every case that the module needs to handle.

Most of the refactoring was done by assuming that the function is_compatible works correctly and ignoring a bunch of possible combinations between source and destination types. For example, a sequential destination dataset can only be used when copying a single file, another sequential dataset or a single member from a PDSE.