Fixed wrong permissions when copying a directory
There was a bug when using the mode option while copying a directory where the parent directory of the destination would end up having the permissions given by the user. This fix makes sure that the module doesn't use the root of the temp directory created to upload local files, resulting in the path returned by copytree (which is used to change permissions) to be the correct directory instead of the parent. A side effect of this change is that whenever a local directory gets copied into a remote system, the module doesn't create another subdirectory in the destination, instead copying the contents into the root of the destination; keep in mind that this is the behavior when copying a remote source.