Ported bugfixes for #664 and #678
* Updated file normalization for sequential datasets

* Fixed encoding when a remote source is used

* Added encoding normalization for record length (#664)

This is a port and tweak from the normalization found in the L2 support branch. Now the normalization applies for both local and remote source files. There is still work left to validate how this behaves when copying to partitioned datasets, and see if the normalization needs to handle directories as well.

* Refactored normalization for USS files

* Updated line endings normalization in PDSEs

* Updated create_temp_with_lf_endings

* Added tests for line endings normalization

* Added changelog fragment